+++ inst/examples/polished_example_01/server.R
@@ -1,7 +1,7 @@
 server <- function(input, output, session) {
 
   output$secure_content <- renderPrint({
-    session$userData$user()
+    session$user
   })